code/requirements.txt
flask
flask-restful
flask-cors
mongoengine
python-dotenv
flasgger
yfinance
redis
rq
flask-jwt-extended
gunicorn
scikit-learn
mlflow
numpy
torch